Move transformable to a separate file
