+++ seismiqb/src/utils/__init__.py
@@ -2,4 +2,5 @@
 # pylint: disable=wildcard-import
 from .classes import *
 from .functions import *
+from .decorators import *
 from .layers import *
